Code for milestone 1.
Changes by Richard Vu:
Update on that second bug: 
- it wasn't due to the double "helward-packard"
- after looking at the tracing, it was not including the last document
in the search
- the way it was before, the pointers for the doc arrays would point to
the last values and get stopped by the while condition, therefore
stopping the  last two from being compared
- I changed it so the while loop sets up the initialization of the
pointers, which has else conditions for advancing the pointers. This way
it does the advancing first, then compares so that it does not get
caught by the while conditional.

- Got a bug where the "or" is printing out duplicates (e.g. inputted
super + king on the hello.txt files). Will try and figure it out, but it
might just be due to the hello files.
- Removed commented out code
- Removed most console print outs
- Removed gloTermLength from SearchEngine
- Removed isLoop, smallSize, and counter (lines 186-193) from
ProcessQuery since they weren't being used for anything
- Left in the code for printResults() and printStatistics()